Vel | Refactor & move wikisource related code to different package
